requirements.txt
fastapi
uvicorn
jinja2
python-multipart
starlette
psycopg2-binary
werkzeug
itsdangerous